[Added] Laravel Authorization

+++ blog/app/Http/Controllers/ArticleController.php
@@ -5,9 +5,16 @@ namespace App\Http\Controllers;
 use App\Models\Article;
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ArticleController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware("auth")->except(['index', 'detail']);
+    }
+
     public function index()
     {
         $data = Article::latest()->paginate(5);
@@ -51,6 +58,7 @@ class ArticleController extends Controller
         $article->title = request()->title;
         $article->body = request()->body;
         $article->category_id = request()->category_id;
+        $article->user_id = auth()->user()->id;
         $article->save();
 
         return redirect('/articles');
@@ -93,8 +101,12 @@ class ArticleController extends Controller
     public function delete($id)
     {
         $article = Article::find($id);
-        $article->delete();
 
-        return redirect('/articles')->with('info', 'An article deleted');
+        if (Gate::allows('delete-article', $article)) {
+            $article->delete();
+            return redirect('/articles')->with('info', 'An article deleted');
+        }
+
+        return back()->with('info', 'Unauthorize to delete');
     }
 }
