+++ blog/app/Http/Controllers/CommentController.php
@@ -4,16 +4,24 @@ namespace App\Http\Controllers;
 
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CommentController extends Controller
 {
+
+    public function __construct()
+    {
+        $this->middleware("auth");
+    }
+
     public function create(Request $request)
     {
         $request->validate(["content" => "required"]);
 
         $comment = new Comment;
-        $comment->content = request()->content;
+        $comment->content = $request->content;
         $comment->article_id = $request->article_id;
+        $comment->user_id = auth()->user()->id;
         $comment->save();
 
         return  back();
@@ -22,8 +30,13 @@ class CommentController extends Controller
     public function delete($id)
     {
         $comment = Comment::find($id);
-        $comment->delete();
 
-        return back()->with('info', 'A comment deleted');
+        if (Gate::allows('delete-comment', $comment)) {
+            $comment->delete();
+            return back()->with('info', 'A comment deleted');
+        }
+
+
+        return back()->with('info', 'Unauthorize to delete');
     }
 }
